move api request to shared utils
